P3/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
nltk>=3.8.0
tqdm>=4.65.0
scikit-learn>=1.3.0

